requirements.txt
Pillow==10.4.0
requests>=2.32.3
python-dotenv>=1.0.1
tqdm>=4.66.4
